The stn_factory and stp_solver_factory are initialized in config.py and imported in stp.py
